+++ frontend/src/components/settings/models/format.js
@@ -1,0 +1,17 @@
+/** Pure, closure-free formatting helpers for the model store. */
+
+export function fmtBytes(n) {
+  if (n == null || n < 0) return '—';
+  if (n === 0) return '0 B';
+  if (n >= 1024 ** 3) return `${(n / 1024 ** 3).toFixed(2)} GB`;
+  if (n >= 1024 ** 2) return `${(n / 1024 ** 2).toFixed(1)} MB`;
+  return `${Math.round(n / 1024)} KB`;
+}
+
+/** Deterministic muted HSL color from an org/user name in a repo_id. */
+export function orgColor(repoId) {
+  const org = (repoId || '').split('/')[0];
+  let h = 0;
+  for (let i = 0; i < org.length; i++) h = (h * 31 + org.charCodeAt(i)) & 0xffff;
+  return `hsl(${h % 360}, 35%, 28%)`;
+}
