+++ frontend/src/pages/Settings.jsx
@@ -3,20 +3,19 @@ import { copyText } from "../utils/copyText";
 import { normalizeChannel } from '../utils/updateChannel';
 import { setChannel } from '../utils/channelControl';
 import {
-  Cpu, FileText, Info, ShieldCheck, RefreshCw, Trash2, ExternalLink,
-  CheckCircle, AlertCircle, Plug, Download, Copy, Building2, KeyRound,
-  Keyboard, Wifi, Palette, Activity, ArrowDownToLine, Settings2,
+  Cpu, FileText, Info, ShieldCheck, RefreshCw,
+  CheckCircle, Plug, KeyRound,
+  Keyboard, Wifi, Palette, ArrowDownToLine, Settings2,
 } from 'lucide-react';
 import { toast } from 'react-hot-toast';
-import { openExternal } from '../api/external';
 import { API } from '../api/client';
-import { Trans, useTranslation } from 'react-i18next';
+import { useTranslation } from 'react-i18next';
 import { systemLogs, systemLogsTauri, clearSystemLogs, clearTauriLogs } from '../api/system';
 import { useSysinfo, useModelStatus, useSystemInfo } from '../api/hooks';
 import { getFrontendLogs, clearFrontendLogs } from '../utils/consoleBuffer';
 import { resolveAboutVersion } from '../utils/appVersion';
-import { Tabs, Segmented, Button, Badge } from '../ui';
-import { SettingsSection, SettingRow } from '../components/settings/primitives';
+import { Tabs, Badge } from '../ui';
+import { SettingsSection } from '../components/settings/primitives';
 import { useAppStore } from '../store';
 import PerformancePanel from '../components/settings/PerformancePanel';
 import RefinementPanel from '../components/settings/RefinementPanel';
@@ -31,12 +30,14 @@ import MCPBindingsPanel from '../components/settings/MCPBindingsPanel';
 import PronunciationPanel from '../components/settings/PronunciationPanel';
 import DictationDemo from '../components/DictationDemo';
 import UpdatesPanel from '../components/UpdatesPanel';
-import ReportBugButton from '../components/ReportBugButton';
 import GeneralTab from '../components/settings/GeneralTab';
 import ModelStoreTab from '../components/settings/ModelStoreTab';
 import EnginesTab from '../components/settings/EnginesTab';
 import HotkeyTab from '../components/settings/HotkeyTab';
 import CredentialsTab from '../components/settings/CredentialsTab';
+import AboutTab from '../components/settings/AboutTab';
+import PrivacyTab from '../components/settings/PrivacyTab';
+import LogsTab from '../components/settings/LogsTab';
 import { isTauri, askConfirm } from '../components/settings/native';
 import './Settings.css';
 
@@ -56,23 +57,6 @@ const TAB_DEFS = [
   { id: 'about',       icon: Info },
   { id: 'privacy',     icon: ShieldCheck },
 ];
-
-const LOG_SOURCE_DEFS = [
-  { value: 'backend',  key: 'backend' },
-  { value: 'frontend', key: 'frontend' },
-  { value: 'tauri',    key: 'tauri' },
-];
-
-
-
-// About/Privacy read-only data rows delegate to the shared SettingRow primitive
-// so they pick up the redesigned grid + mono value styling unchanged.
-function Row({ label, value, mono }) {
-  return <SettingRow title={label} control={value} mono={mono} />;
-}
-
-
-
 
 export default function Settings() {
   const { t } = useTranslation();
@@ -370,58 +354,15 @@ export default function Settings() {
       {activeTab === 'credentials' && <CredentialsTab info={info} />}
 
       {activeTab === 'logs' && (
-        <SettingsSection
-          icon={FileText}
-          title={t('settings.logs')}
-          actions={
-            <>
-              <ReportBugButton />
-              <Button
-                variant="subtle"
-                size="sm"
-                onClick={refreshLogs}
-                loading={loadingLogs}
-                leading={!loadingLogs && <RefreshCw size={11} />}
-              >
-                {t('common.refresh')}
-              </Button>
-              <Button
-                variant="danger"
-                size="sm"
-                onClick={onClearLogs}
-                leading={<Trash2 size={11} />}
-              >
-                {t('common.clear')}
-              </Button>
-            </>
-          }
-        >
-          <Segmented
-            items={LOG_SOURCE_DEFS.map(d => ({ ...d, label: t(`common.${d.key}`) }))}
-            value={logSource}
-            onChange={setLogSource}
-          />
-
-          <div className="settings-log-meta">
-            <span>{logMeta.path || '—'}</span>
-            {logSource === 'tauri' && !logMeta.exists && (
-              <Badge tone="warn">
-                <AlertCircle size={11} /> {t('logs.no_tauri_log')}
-              </Badge>
-            )}
-          </div>
-          <div className="settings-log">
-            {logs.length === 0
-              ? <span className="settings-log__empty">
-                  {logSource === 'frontend'
-                    ? t('logs.empty_frontend')
-                    : logSource === 'tauri'
-                      ? t('logs.empty_tauri')
-                      : t('logs.empty_backend')}
-                </span>
-              : logs.join('')}
-          </div>
-        </SettingsSection>
+        <LogsTab
+          logSource={logSource}
+          setLogSource={setLogSource}
+          logs={logs}
+          logMeta={logMeta}
+          loadingLogs={loadingLogs}
+          refreshLogs={refreshLogs}
+          onClearLogs={onClearLogs}
+        />
       )}
 
       {activeTab === 'updates' && (
@@ -431,170 +372,26 @@ export default function Settings() {
       )}
 
       {activeTab === 'about' && (
-        <SettingsSection icon={Info} title={t('settings.about')}>
-          <Row label={t('about.app')}             value="OmniVoice Studio" />
-          <Row label={t('about.version')}         value={resolveAboutVersion(appVersion, info)} mono />
-          <Row label={t('about.tauri_runtime')}   value={tauriVersion || (isTauri() ? '—' : t('about.web_preview'))} mono />
-          <Row label={t('about.platform')}        value={info?.platform || '—'} />
-          <Row label={t('about.architecture')}    value={info?.arch || '—'} mono />
-          <Row label={t('about.python')}          value={info?.python || '—'} mono />
-          <Row label={t('about.compute_device')}  value={info?.device || '—'} mono />
-          <Row label={t('about.gpu_active')}      value={hw?.gpu_active
-            ? <Badge tone="success"><CheckCircle size={11} /> {t('about.yes')}</Badge>
-            : <Badge tone="neutral">{t('about.no')}</Badge>} />
-          <Row label={t('about.ram')}             value={hw ? `${hw.ram?.toFixed(2)} / ${hw.total_ram?.toFixed(2)} GB` : '—'} mono />
-          <Row label={t('about.vram')}            value={hw ? `${hw.vram?.toFixed(2)} GB` : '—'} mono />
-          <Row label={t('about.backend')}         value={<Badge tone={status?.status === 'ready' ? 'success' : status?.status === 'loading' ? 'warn' : 'neutral'}>{status?.status || 'unknown'}</Badge>} />
-          <Row label={t('about.active_model')}    value={status?.repo_id || info?.model_checkpoint || '—'} mono />
-          <Row label={t('about.asr_model')}       value={info?.asr_model || '—'} mono />
-          <Row label={t('about.translator')}      value={info?.translate_provider || '—'} />
-          <Row label={t('about.hf_token')}        value={info?.has_hf_token ? t('about.yes') : t('about.no')} />
-          <Row label={t('about.data_dir')}        value={info?.data_dir || '—'} mono />
-          <Row label={t('about.outputs')}         value={info?.outputs_dir || '—'} mono />
-          <Row label={t('about.crash_log')}       value={info?.crash_log_path || '—'} mono />
-          {/* Auto-updater + channel toggle are desktop-only (Tauri). The Docker
-              web build updates by pulling a new image tag, so hide these rows
-              there to avoid a non-functional control (issue #249). */}
-          {isTauri() && (
-            <>
-              <SettingRow
-                title={t('about.update_channel')}
-                hint={updateChannel === 'preview' ? t('about.channel_preview_hint') : undefined}
-                control={
-                  <Segmented
-                    size="xs"
-                    value={updateChannel}
-                    onChange={changeChannel}
-                    items={[
-                      { value: 'stable',  label: t('about.channel_stable') },
-                      { value: 'preview', label: t('about.channel_preview') },
-                    ]}
-                  />
-                }
-              />
-              <Row
-                label={t('about.update_endpoint')}
-                value={updateChannel === 'preview'
-                  ? 'releases/download/preview/latest.json'
-                  : 'releases/latest/download/latest.json'}
-                mono
-              />
-            </>
-          )}
-          <div className="settings-link-row">
-            {isTauri() && (
-              <Button
-                variant="primary"
-                size="md"
-                leading={<Download size={12} />}
-                onClick={checkForUpdates}
-                loading={updateState === 'checking' || updateState === 'downloading'}
-              >
-                {updateState === 'downloading' ? t('about.downloading') : t('about.check_updates')}
-              </Button>
-            )}
-            <Button
-              variant="subtle"
-              size="md"
-              leading={!selfCheckRunning && <Activity size={12} />}
-              onClick={runSelfCheck}
-              loading={selfCheckRunning}
-            >
-              {t('about.self_check')}
-            </Button>
-            <Button
-              variant="subtle"
-              size="md"
-              leading={!bundleBuilding && <Download size={12} />}
-              onClick={saveDiagnosticBundle}
-              loading={bundleBuilding}
-            >
-              {t('about.save_bundle')}
-            </Button>
-            <Button
-              variant="subtle"
-              size="md"
-              leading={<Copy size={12} />}
-              onClick={copyDiagnostics}
-            >
-              {t('about.copy_diagnostics')}
-            </Button>
-            <Button
-              variant="subtle"
-              size="md"
-              leading={<ExternalLink size={12} />}
-              onClick={() => openExternal('https://github.com/k2-fsa/OmniVoice')}
-            >
-              {t('about.github')}
-            </Button>
-            <Button
-              variant="subtle"
-              size="md"
-              leading={<ExternalLink size={12} />}
-              onClick={() => openExternal('https://huggingface.co/k2-fsa/OmniVoice')}
-            >
-              {t('about.model_card')}
-            </Button>
-            <Button
-              variant="subtle"
-              size="md"
-              leading={<Building2 size={12} />}
-              onClick={() => { useAppStore.getState().setMode?.('enterprise'); }}
-            >
-              {t('about.commercial_license')}
-            </Button>
-          </div>
-          {selfCheck && (
-            <div className="settings-selfcheck">
-              {selfCheck.checks.map((c) => (
-                <Row
-                  key={c.id}
-                  label={c.label}
-                  value={
-                    <span>
-                      <Badge tone={c.status === 'ok' ? 'success' : c.status === 'warn' ? 'warn' : 'danger'}>
-                        {c.status === 'ok'
-                          ? <CheckCircle size={11} />
-                          : <AlertCircle size={11} />} {t(`about.self_check_${c.status}`)}
-                      </Badge>
-                      {' '}{c.detail}
-                      {c.hint && <span className="settings-muted"> — {c.hint}</span>}
-                    </span>
-                  }
-                />
-              ))}
-              <p className="settings-muted">
-                {selfCheck.summary.ok
-                  ? t('about.self_check_healthy')
-                  : t('about.self_check_attention', { count: selfCheck.summary.failures })}
-              </p>
-            </div>
-          )}
-        </SettingsSection>
+        <AboutTab
+          appVersion={appVersion}
+          tauriVersion={tauriVersion}
+          info={info}
+          hw={hw}
+          status={status}
+          updateChannel={updateChannel}
+          changeChannel={changeChannel}
+          checkForUpdates={checkForUpdates}
+          updateState={updateState}
+          selfCheck={selfCheck}
+          selfCheckRunning={selfCheckRunning}
+          runSelfCheck={runSelfCheck}
+          bundleBuilding={bundleBuilding}
+          saveDiagnosticBundle={saveDiagnosticBundle}
+          copyDiagnostics={copyDiagnostics}
+        />
       )}
 
-      {activeTab === 'privacy' && (
-        <SettingsSection icon={ShieldCheck} title={t('settings.privacy')}>
-          <p className="settings-prose">
-            <Trans i18nKey="privacy.desc" components={{ 1: <strong /> }} />
-          </p>
-          <Row label={t('privacy.uploads_at')}   value={info?.data_dir ? `${info.data_dir}/` : '—'} mono />
-          <Row label={t('privacy.outputs_at')}   value={info?.outputs_dir || '—'} mono />
-          <Row label={t('privacy.gen_history')}  value={<Badge tone="neutral">{t('privacy.local_sqlite')}</Badge>} />
-          <Row
-            label={t('privacy.network_calls')}
-            value={
-              info?.translate_provider && ['google', 'deepl', 'mymemory', 'microsoft', 'openai'].includes(info.translate_provider)
-                ? <Badge tone="warn"><AlertCircle size={11} /> {t('privacy.translator_online', { provider: info.translate_provider })}</Badge>
-                : <Badge tone="success"><CheckCircle size={11} /> {t('privacy.translator_offline')}</Badge>
-            }
-          />
-          <Row
-            label={t('privacy.model_telemetry')}
-            value={<Badge tone="success"><CheckCircle size={11} /> {t('privacy.no_tracking')}</Badge>}
-          />
-        </SettingsSection>
-      )}
+      {activeTab === 'privacy' && <PrivacyTab info={info} />}
       </div>
     </div>
   );
